// web/src/api/models/default-model-config.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DefaultModelConfig
 */
export interface DefaultModelConfig {
    /**
     * The scenario for which this default model is configured
     * @type {string}
     * @memberof DefaultModelConfig
     */
    'scenario': DefaultModelConfigScenarioEnum;
    /**
     * 
     * @type {string}
     * @memberof DefaultModelConfig
     */
    'custom_llm_provider'?: string;
    /**
     * The name of the model provider
     * @type {string}
     * @memberof DefaultModelConfig
     */
    'provider_name'?: string;
    /**
     * The name of the model
     * @type {string}
     * @memberof DefaultModelConfig
     */
    'model'?: string;
}

export const DefaultModelConfigScenarioEnum = {
    default_for_collection_completion: 'default_for_collection_completion',
    default_for_agent_completion: 'default_for_agent_completion',
    default_for_embedding: 'default_for_embedding',
    default_for_rerank: 'default_for_rerank',
    default_for_background_task: 'default_for_background_task'
} as const;

export type DefaultModelConfigScenarioEnum = typeof DefaultModelConfigScenarioEnum[keyof typeof DefaultModelConfigScenarioEnum];